Redis refactoring
split parser and unexport/rename some types
